lambda/requirements.txt
python-magic==0.4.27
boto3==1.34.69
requests>=2.31.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-docx>=0.8.11
